\name{topicar-package}
\alias{topicar-package}
\alias{topicar}
\docType{package}
\title{
\packageTitle{topicar}
}
\description{
\packageDescription{topicar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{topicar}
\packageIndices{topicar}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{topicar}

Maintainer: \packageMaintainer{topicar}
}
\references{
none
}
%%~~ Optionally other standard keywords, one per line, from file ~~
%%~~ KEYWORDS in the R documentation directory ~~
%%\keyword{ package }
%%\seealso{
%%~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}
%%\examples{
%%~~ simple examples of the most important functions ~~
%%}
